fixed bug on [ index_ddl_error.py, network_write_cond_wakeup.py, og_easy_slow.py] (#1073)
* Update index_ddl_error.py

There is a bug here. If event_data is not None but empty

* Update index_ddl_error.py

* Update log_easy_slow.py

There is a bug here. 

If 'EASY SLOW' is not found in the log file, 'grep' fails and then 'echo' is triggered, the result will be '0\n0'.

Then, 'easy_slow_count = int(result)' failed

* Update network_write_cond_wakeup.py

There is a bug here. 

If 'write cond wakeup' is not found in the log file, 'grep' fails and then 'echo' is triggered, the result will be '0\n0'.

Then, 'easy_slow_count = int(result)' failed

---------

Co-authored-by: xuyan wang <35394786+wayyoungboy@users.noreply.github.com>