Missing required command-line arguments in this file [src/common/diag_cmd.py] (#1078)
* Update index_ddl_error.py

There is a bug here. If event_data is not None but empty

* Update index_ddl_error.py

* Update log_easy_slow.py

There is a bug here. 

If 'EASY SLOW' is not found in the log file, 'grep' fails and then 'echo' is triggered, the result will be '0\n0'.

Then, 'easy_slow_count = int(result)' failed

* Update network_write_cond_wakeup.py

There is a bug here. 

If 'write cond wakeup' is not found in the log file, 'grep' fails and then 'echo' is triggered, the result will be '0\n0'.

Then, 'easy_slow_count = int(result)' failed

* Update ddl_failure.py

This query did not find any records. It should not prompt that the table was not found, as it may cause confusion among users

* Update diag_cmd.py

Missing required command-line arguments

---------

Co-authored-by: xuyan wang <35394786+wayyoungboy@users.noreply.github.com>