Failed to obtain the $LANG variable of the observer node (#1085)
* Update index_ddl_error.py

There is a bug here. If event_data is not None but empty

* Update index_ddl_error.py

* Update log_easy_slow.py

There is a bug here. 

If 'EASY SLOW' is not found in the log file, 'grep' fails and then 'echo' is triggered, the result will be '0\n0'.

Then, 'easy_slow_count = int(result)' failed

* Update network_write_cond_wakeup.py

There is a bug here. 

If 'write cond wakeup' is not found in the log file, 'grep' fails and then 'echo' is triggered, the result will be '0\n0'.

Then, 'easy_slow_count = int(result)' failed

* Update ddl_failure.py

This query did not find any records. It should not prompt that the table was not found, as it may cause confusion among users

* Update diag_cmd.py

Missing required command-line arguments

* Update check_system_language.py

Here are three issues:
1.Directly using $LANG retrieves the variable from the local machine. To obtain the variable value from the SSH target machine, the variable must be escaped or enclosed in single quotes.  [For example : echo '${LANG}' or echo \${LANG}]
2.On some systems, ssh may not load environment variables, causing the retrieved variable value to differ from the one obtained when logging directly into the machine (this issue has been observed on Debian 9).
3.Some systems return uppercase values while others return lowercase, leading to inconsistent formats. The solution needs to accommodate multiple formats.[For example : Rocky Linux release 9.5 (Blue Onyx)]

---------

Co-authored-by: xuyan wang <35394786+wayyoungboy@users.noreply.github.com>